+++ NEWS.md
@@ -10,6 +10,9 @@
 * mkvinfo: added support for the new track header elements
   `BlockAdditionMapping`, `BlockAddIDType`, `BlockAddIDValue` and
   `BlockAddIDExtraData`. Part of the implementation of #2887.
+* mkvmerge: added support for copying the new track header elements
+  `BlockAddIDType`, `BlockAddIDValue` and `BlockAddIDExtraData` from Matroska
+  files if they exist. Part of the implementation of #2887.
 
 ## Bug fixes